+++ packages/headless/src/utils/errors.ts
@@ -9,9 +9,11 @@ export class ExpiredTokenError extends Error {
 }
 
 export class DisconnectedError extends Error {
-  constructor() {
+  public statusCode: number;
+  constructor(url: string, statusCode?: number) {
     super();
     this.name = 'Disconnected';
-    this.message = 'Client is not connected to the internet.';
+    this.message = `Client could not connect to the following URL: ${url}`;
+    this.statusCode = statusCode ?? 0;
   }
 }